+++ eslint.config.ts
@@ -1,0 +1,12 @@
+import js from '@eslint/js'
+import globals from 'globals'
+import tseslint from 'typescript-eslint'
+
+export default [
+  js.configs.recommended,
+  {
+    files: ['**/*.{js,mjs,cjs,ts,mts,cts}'],
+    languageOptions: { globals: globals.browser },
+  },
+  ...tseslint.configs.recommended,
+]